Stacking bugfix
By testing the d.x and d.y and not the gotoX and gotoY the game allows you to stack many shapes on top of each other. To see this in action try changing lines 172&173 to:
		self.x = self.x*0.9 + self.gotoX*0.1;
		self.y = self.y*0.9 + self.gotoY*0.1;
Then pick up a shape and move it onto a blocked space and let it go, then as it's returning to it's original space move another shape to that original space. This should fix that problem.